move media handling into separate class
