+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/constants.ts
@@ -1,5 +1,3 @@
-export { UMB_DROPZONE_MEDIA_TYPE_PICKER_MODAL } from './modals/dropzone-media-type-picker/dropzone-media-type-picker-modal.token.js';
-
 export enum UmbFileDropzoneItemStatus {
 	WAITING = 'waiting',
 	COMPLETE = 'complete',
