+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/index.ts
@@ -1,7 +1,6 @@
-export * from './constants.js';
 export * from './components/index.js';
-export * from './modals/index.js';
+export * from './constants.js';
+export * from './dropzone-change.event.js';
 export * from './dropzone-manager.class.js';
 export * from './dropzone-submitted.event.js';
-export * from './dropzone-change.event.js';
 export type * from './types.js';
