+++ src/Umbraco.Web.UI.Client/src/packages/media/dropzone/manifests.ts
@@ -1,3 +1,3 @@
-import { manifests as modalManifests } from './modals/manifests.js';
+import { manifests as modalManifests } from '../media/dropzone/modals/manifests.js';
 
 export const manifests: Array<UmbExtensionManifest> = [...modalManifests];
