+++ src/Umbraco.Web.UI.Client/src/packages/media/media/dropzone/modals/index.ts
